add medianofmedian.cpp
Given an array and a number k where k is smaller than the size of the array, we need to find the k’th smallest element in the given array. It is given that all array elements are distinct.At that case this worst case linear time algo can be used.